+++ packages/create-gasket-app/CHANGELOG.md
@@ -1,5 +1,6 @@
 # `create-gasket-app`
 
+- Add `remove` to config-builder ([#1036])
 - Add `addCommand` method for create context ([#1034])
 - Add `addEnvironment` method for create context ([#1010])
 
@@ -11,7 +12,7 @@
 
 - Aligned version releases across all packages
 
-Added `@gasket/plugin-dynamic-plugin` to default plugins ([#970])
+- Added `@gasket/plugin-dynamic-plugin` to default plugins ([#970])
 
 ### 7.0.6
 
@@ -94,3 +95,4 @@ Added `@gasket/plugin-dynamic-plugin` to default plugins ([#970])
 [#1010]: https://github.com/godaddy/gasket/pull/1010
 [#1014]: https://github.com/godaddy/gasket/pull/1014
 [#1034]: https://github.com/godaddy/gasket/pull/1034
+[#1036]: https://github.com/godaddy/gasket/pull/1036